fix: path to config file
A project can be run from any folder, so there is no guarantee that Directory.GetCurrentFolder is anywhere near the project folder. This makes it hard to find the appsettings file.
This is a known problem (see https://github.com/dotnet/cli/issues/7316).
For now, make sure everything works if the project is run from one of the two most likely places: the solution folder, or the project folder.